% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddqc.R
\name{initialQC}
\alias{initialQC}
\title{Perform initial quality control}
\usage{
initialQC(
  data,
  basic.n.genes = 100,
  basic.percent.mt = 80,
  mt.prefix = "MT-",
  rb.prefix = "^RP[SL][[:digit:]]|^RPLP[[:digit:]]|^RPSA"
)
}
\arguments{
\item{data}{Seurat object}

\item{basic.n.genes}{lower nFeature_RNA filtering. 100 by default}

\item{basic.percent.mt}{upper percent.mt filtering. 80 by default}

\item{mt.prefix}{gene regular expression used to calculate percent.mt in a cell
"MT-" by default}

\item{rb.prefix}{gene regular expression used to calculate percent.rb in a cell}
}
\value{
Filltered Seurat object
}
\description{
This function takes a Seurat object and does initial QC on it
}
