% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddqc.R
\name{ddqc.metrics}
\alias{ddqc.metrics}
\title{Calculate which cells are passing ddqc}
\usage{
ddqc.metrics(
  data,
  res = 1,
  threshold = 2,
  do.counts = TRUE,
  do.genes = TRUE,
  do.mito = TRUE,
  do.ribo = FALSE,
  n.genes.lower.bound = 200,
  percent.mito.upper.bound = 10,
  random.state = 29
)
}
\arguments{
\item{data}{Seurat object}

\item{res}{clustering resolution. 1 by default}

\item{threshold}{MAD multiplier for ddqc. 2 by default}

\item{do.counts}{whether to consider nCount_RNA for ddqc. TRUE by default}

\item{do.genes}{whether to consider nFeature_RNA for ddqc. TRUE by default}

\item{do.mito}{whether to consider percent.mt for ddqc. TRUE by default}

\item{do.ribo}{whether to consider percent.rb for ddqc. TRUE by default}

\item{n.genes.lower.bound}{bound for lower nFeature_RNA cluster-level threshold. 200 by default}

\item{percent.mito.upper.bound}{bound for upper percent.mt cluster-level threshold. 10 by default}

\item{random.state}{random seed for clustering results reproducibility. 29 by default}
}
\value{
data.frame with ddqc statistics
}
\description{
This function takes a Seurat object after InitialQC, and then performs ddqc on it
Returns a data.frame that tells which cells have passed ddqc and additional information
}
